[#107] runtests.bat: Allow user to run script outside test folder (#108)
runtests.bat script does not work outside the test folder.

Developers working on the project often have their working directory
set to the root of the project. It is inconvenient to switch in and out
of the test folder repeatedly in an edit-debug cycle.

To reduce the hassle, let's add the commands in bat script to perform
the switching of directory automatically and hence increase our
productivity.