DeveloperGuide: add solution to src being treated as normal folder (#105)
Occasionally, new developers may accidentally import the project as
a Gradle project.

This cause IntelliJ to often triggers Gradle synchronization which
allows Gradle to impose the project to use its configuration. Due
to this issue, developers will encounter the problem of marked
directory keep getting reset upon restarting IntelliJ.

Let's add the solution to this problem in the troubleshooting
section to help the developers who have accidentally misconfigured
the importation.